chore: renames SQL SDK package to be coherent with other SDK packages.
