+++ sdk/database/sql/sql_test.go
@@ -1,4 +1,4 @@
-package hypersql
+package sql
 
 // highly inspired in https://github.com/openzipkin-contrib/zipkin-go-sql/blob/master/driver_test.go
